+++ src/PropertyValidator/IExpressPropertyValidator.cs
@@ -1,4 +1,4 @@
 ﻿namespace ExpressValidator
 {
-	internal interface IExpressPropertyValidator<T> :  IExpressPropertyValidatorBase<T>, IObjectValidator{}
+	internal interface IExpressPropertyValidator<TObj, T> :  IExpressPropertyValidatorBase<T>, IObjectValidator<TObj> { }
 }