+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.14
+
+- Introduce `ExpressValidatorBuilder.AddFunc` method.
+- Update to FluentValidation 11.9.1.
+- Made the `BuilderWithPropValidator` constructor internal.
+- Add `TObj` generic param to `IExpressPropertyValidator`, `ExpressPropertyValidator`.
+
+
 ## 0.0.10
 
 - Support for adding a field for validation - introduced `ExpressValidatorBuilder.AddField` method.